add localized attributes for subscription period and introductory period to SKProduct
